Footer refactor for vanilla store
