+++ templates/demo-store/src/components/global/Footer.server.tsx
@@ -1,0 +1,42 @@
+import {useUrl} from '@shopify/hydrogen';
+
+import {Section, Heading, FooterMenu, CountrySelector} from '~/components';
+import type {EnhancedMenu} from '~/lib/utils';
+
+/**
+ * A server component that specifies the content of the footer on the website
+ */
+export function Footer({menu}: {menu?: EnhancedMenu}) {
+  const {pathname} = useUrl();
+
+  const localeMatch = /^\/([a-z]{2})(\/|$)/i.exec(pathname);
+  const countryCode = localeMatch ? localeMatch[1] : null;
+
+  const isHome = pathname === `/${countryCode ? countryCode + '/' : ''}`;
+  const itemsCounts = menu ? menu?.items?.length + 1 : [];
+
+  return (
+    <Section
+      divider={isHome ? 'none' : 'top'}
+      as="footer"
+      role="contentinfo"
+      className={[
+        'grid items-start grid-flow-row w-full gap-6',
+        `py-8 px-6 md:px-8 lg:px-12 border-b md:gap-8 grid-cols-1 lg:gap-12 md:grid-cols-2 lg:grid-cols-${itemsCounts}`,
+        'bg-primary dark:bg-contrast dark:text-primary text-contrast',
+      ].join(' ')}
+    >
+      <FooterMenu menu={menu} />
+      <section className="grid gap-4 w-full md:max-w-[335px] md:ml-auto">
+        <Heading size="lead" className="cursor-default" as="h4">
+          Country
+        </Heading>
+        <CountrySelector />
+      </section>
+      <div className="pt-8 opacity-50 md:col-span-2 lg:col-span-4">
+        &copy; {new Date().getFullYear()} / Shopify, Inc. Hydrogen is an MIT
+        Licensed Open Source project. This website is carbon&nbsp;neutral.
+      </div>
+    </Section>
+  );
+}
