+++ templates/demo-store/src/components/global/FooterMenu.client.tsx
@@ -1,0 +1,55 @@
+// @ts-expect-error @headlessui/react incompatibility with node16 resolution
+import {Disclosure} from '@headlessui/react';
+import {Link, useUrl} from '@shopify/hydrogen';
+
+import {Heading, IconCaret} from '~/components';
+import type {EnhancedMenu, EnhancedMenuItem} from '~/lib/utils';
+
+/**
+ * A server component that specifies the content of the footer on the website
+ */
+export function FooterMenu({menu}: {menu?: EnhancedMenu}) {
+  const styles = {
+    section: 'grid gap-4',
+    nav: 'grid gap-2 pb-6',
+  };
+
+  return (menu?.items || []).map((item: EnhancedMenuItem) => (
+    <section key={item.id} className={styles.section}>
+      <Disclosure>
+        {/* @ts-expect-error @headlessui/react incompatibility with node16 resolution */}
+        {({open}) => (
+          <>
+            <Disclosure.Button className="text-left md:cursor-default">
+              <Heading className="flex justify-between" size="lead" as="h4">
+                {item.title}
+                <span className="md:hidden">
+                  <IconCaret direction={open ? 'up' : 'down'} />
+                </span>
+              </Heading>
+            </Disclosure.Button>
+            <div
+              className={`${
+                open ? `max-h-48 h-fit` : `max-h-0 md:max-h-fit`
+              } overflow-hidden transition-all duration-300`}
+            >
+              <Disclosure.Panel static>
+                <nav className={styles.nav}>
+                  {item.items.map((subItem) => (
+                    <Link
+                      key={subItem.id}
+                      to={subItem.to}
+                      target={subItem.target}
+                    >
+                      {subItem.title}
+                    </Link>
+                  ))}
+                </nav>
+              </Disclosure.Panel>
+            </div>
+          </>
+        )}
+      </Disclosure>
+    </section>
+  ));
+}
