+++ templates/demo-store/src/components/global/index.server.ts
@@ -1,2 +1,3 @@
-export {NotFound} from './NotFound.server';
+export {Footer} from './Footer.server';
 export {Layout} from './Layout.server';
+export {NotFound} from './NotFound.server';
