+++ templates/demo-store/src/components/global/index.ts
@@ -1,5 +1,5 @@
 export {Drawer, useDrawer} from './Drawer.client';
-export {Footer} from './Footer.client';
+export {FooterMenu} from './FooterMenu.client';
 export {Header} from './Header.client';
 export {Modal} from './Modal.client';
 export {PageHeader} from './PageHeader';
